app/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.21.1
PyJWT==2.8.0
pydantic==1.10.11
pytest==7.4.0
pytest-cov==4.1.0
httpx==0.24.1
